requirements.txt
pyyaml==6.0.2
pyspark==3.4.1
dotenv==0.9.9
pandas==2.3.0
numpy < 2.0.0
requests==2.32.4
papermill==2.6.0
